Proxy config fixup for bad merge.
The previous proxy config patch erroneously applied changes to the deleted
file package_managers.py, reviving it in the process. Delete it again, and
apply its changes to linux_virtual_machine.py.
